feat: refactor Table [MDS-610]
