+++ next-docs/public/examples/table/ExpandedRow.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import { Table } from '@heathmont/moon-table-tw';
-import { BodyTR } from '@heathmont/moon-table-tw';
-import { TD } from '@heathmont/moon-table-tw';
 
 interface HeaderProps {
   isAllRowsExpanded: boolean;
