+++ workspaces/tables/src/components/Body.tsx
@@ -1,12 +1,6 @@
-import React, { forwardRef, ReactNode } from 'react';
+import React, { forwardRef } from 'react';
 import { mergeClassnames } from '@heathmont/moon-core-tw';
-import type RowGaps from '../private/types/RowGaps';
-
-type BodyProps = {
-  children?: ReactNode;
-  reactTableProps?: any;
-  rowGap?: RowGaps;
-};
+import type BodyProps from '../private/types/BodyProps';
 
 const Body = forwardRef<HTMLDivElement, BodyProps>(
   ({ children, reactTableProps, rowGap }) => (
