+++ workspaces/tables/src/components/CheckboxTD.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const CheckboxTD: React.FC<{ children?: React.ReactNode }> = ({ children }) => (
+const CheckboxTD = ({ children }: { children?: React.ReactNode }) => (
   <div className="flex items-center h-full w-full justify-center pl-2">
     {children}
   </div>
