+++ workspaces/tables/src/components/Footer.tsx
@@ -1,29 +1,21 @@
 import React from 'react';
+import type FooterProps from '../private/types/FooterProps';
 
-type FooterProps = {
-  ref: React.RefObject<any>;
-  headerBackgroundColor: string;
-  selectable?: boolean;
-  children?: React.ReactNode;
-};
-
-const Footer: React.FC<FooterProps> = ({
+const Footer = ({
   headerBackgroundColor,
   selectable,
   children,
-}) => {
-  return (
-    <div
-      className={'sticky z-[99] bottom-0'}
-      style={
-        selectable
-          ? { backgroundColor: `rgb(var(--${headerBackgroundColor}))` }
-          : {}
-      }
-    >
-      {children}
-    </div>
-  );
-};
+}: FooterProps) => (
+  <div
+    className="sticky z-[99] bottom-0"
+    style={
+      selectable
+        ? { backgroundColor: `rgb(var(--${headerBackgroundColor}))` }
+        : {}
+    }
+  >
+    {children}
+  </div>
+);
 
 export default Footer;
