+++ workspaces/tables/src/components/Header.tsx
@@ -1,28 +1,21 @@
 import React from 'react';
+import type HeaderProps from '../private/types/HeaderProps';
 
-type HeaderProps = {
-  headerBackgroundColor: string;
-  selectable?: boolean;
-  children?: React.ReactNode;
-};
-
-const Header: React.FC<HeaderProps> = ({
+const Header = ({
   headerBackgroundColor,
   selectable,
   children,
-}) => {
-  return (
-    <div
-      className={'flex flex-col sticky top-0 z-[99]'}
-      style={
-        selectable
-          ? { backgroundColor: `rgb(var(--${headerBackgroundColor}))` }
-          : {}
-      }
-    >
-      {children}
-    </div>
-  );
-};
+}: HeaderProps) => (
+  <div
+    className="flex flex-col sticky top-0 z-[99]"
+    style={
+      selectable
+        ? { backgroundColor: `rgb(var(--${headerBackgroundColor}))` }
+        : {}
+    }
+  >
+    {children}
+  </div>
+);
 
 export default Header;
