+++ workspaces/tables/src/components/HeaderTR.tsx
@@ -1,18 +1,12 @@
 import React, { forwardRef } from 'react';
-
-type HeaderTRProps = {
-  children: any;
-  reactTableProps?: any;
-};
+import type HeaderTRProps from '../private/types/HeaderTRProps';
 
 const HeaderTR = forwardRef<HTMLDivElement, HeaderTRProps>(
-  ({ reactTableProps, children }) => {
-    return (
-      <div {...reactTableProps} className="flex rounded-moon-s-sm sticky">
-        {children}
-      </div>
-    );
-  }
+  ({ reactTableProps, children }) => (
+    <div {...reactTableProps} className="flex rounded-moon-s-sm sticky">
+      {children}
+    </div>
+  )
 );
 
 export default HeaderTR;
