+++ workspaces/tables/src/components/OuterWrapper.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 
-const OuterWrapper: React.FC<{ children?: React.ReactNode }> = ({
-  children,
-}) => {
-  return <div className="relative h-full w-full overflow-auto">{children}</div>;
-};
+const OuterWrapper = ({ children }: { children?: React.ReactNode }) => (
+  <div className="relative h-full w-full overflow-auto">{children}</div>
+);
 
 export default OuterWrapper;
