+++ workspaces/tables/src/components/TD.tsx
@@ -1,32 +1,9 @@
-import React, { forwardRef, ReactNode } from 'react';
+import React, { forwardRef } from 'react';
 import { mergeClassnames } from '@heathmont/moon-core-tw';
-import { TableVariant } from '../func/Table';
 import getCellBorder from '../private/utils/getCellBorder';
 import getFontSize from '../private/utils/getFontSize';
 import getPadding from '../private/utils/getPadding';
-import type RowSizes from '../private/types/RowSizes';
-
-type TDProps = {
-  reactTableProps?: any;
-  selectable?: boolean;
-  isExpanded?: boolean;
-  isLastRow?: boolean;
-  hasParent?: boolean;
-  isSelected?: boolean | any;
-  customBackground?: boolean;
-  backgroundColor?: string;
-  fontColor?: string;
-  children?: ReactNode;
-  stickySide?: string;
-  isFirstColumn?: boolean;
-  isLastColumn?: boolean;
-  isHovered?: boolean;
-  variant?: TableVariant;
-  headerBackgroundColor?: string;
-  onClick?: () => void;
-  rowSize?: RowSizes;
-  isCellBorder?: boolean;
-};
+import type TDProps from '../private/types/TDProps';
 
 const TD = forwardRef<HTMLDivElement, TDProps>(
   ({
@@ -53,8 +30,8 @@ const TD = forwardRef<HTMLDivElement, TDProps>(
       {...reactTableProps}
       className={mergeClassnames(
         'relative box-border justify-between items-center text-start',
-        getFontSize(rowSize as RowSizes),
-        getPadding(rowSize as RowSizes),
+        getFontSize(rowSize),
+        getPadding(rowSize),
         isFirstColumn && 'rounded-l-lg',
         isLastColumn ? 'rounded-r-lg' : getCellBorder(isCellBorder),
         isExpanded && 'first:rounded-bl-lg last:rounded-br-lg after:hidden',
