+++ workspaces/tables/src/components/TH.tsx
@@ -1,21 +1,9 @@
-import React, { forwardRef, ReactNode } from 'react';
+import React, { forwardRef } from 'react';
 import { mergeClassnames } from '@heathmont/moon-core-tw';
-import { TableVariant } from '../func/Table';
 import getCellBorder from '../private/utils/getCellBorder';
 import getFontSize from '../private/utils/getFontSize';
 import getPadding from '../private/utils/getPadding';
-import type RowSizes from '../private/types/RowSizes';
-
-type THProps = {
-  headerBackgroundColor?: string;
-  stickySide?: string;
-  children?: ReactNode;
-  reactTableProps?: any;
-  isLastColumn?: boolean;
-  variant?: TableVariant;
-  rowSize?: RowSizes;
-  isCellBorder?: boolean;
-};
+import type THProps from '../private/types/THProps';
 
 const TH = forwardRef<HTMLDivElement, THProps>(
   ({
@@ -32,8 +20,8 @@ const TH = forwardRef<HTMLDivElement, THProps>(
       {...reactTableProps}
       className={mergeClassnames(
         'relative text-start justify-between font-medium',
-        getFontSize(rowSize as RowSizes),
-        getPadding(rowSize as RowSizes),
+        getFontSize(rowSize),
+        getPadding(rowSize),
         !isLastColumn && getCellBorder(isCellBorder),
         `bg-${headerBackgroundColor}`,
         variant === 'calendar' && 'first:after:hidden',
