+++ workspaces/tables/src/index.ts
@@ -28,13 +28,7 @@ export { default as TD } from './components/TD';
 export * from './components/TD';
 export { default as TH } from './components/TH';
 export * from './components/TH';
+export { default as Table } from './components/Table';
+export * from './components/Table';
 export { default as TableWrapper } from './components/TableWrapper';
 export * from './components/TableWrapper';
-export { default as Table } from './func/Table';
-export * from './func/Table';
-export { default as useRowSpan } from './hooks/useRowSpan';
-export * from './hooks/useRowSpan';
-export { default as useScrollState } from './hooks/useScrollState';
-export * from './hooks/useScrollState';
-export { default as StyledTable } from './styled/StyledTable';
-export * from './styled/StyledTable';
