+++ workspaces/tables/src/private/hooks/useScrollState.ts
