+++ workspaces/tables/src/private/types/BodyProps.ts
@@ -1,0 +1,9 @@
+import type RowGaps from './RowGaps';
+
+type BodyProps = {
+  children?: React.ReactNode;
+  reactTableProps?: any;
+  rowGap?: RowGaps;
+};
+
+export default BodyProps;
