+++ workspaces/tables/src/private/types/FooterProps.ts
@@ -1,0 +1,8 @@
+type FooterProps = {
+  ref: React.RefObject<any>;
+  headerBackgroundColor: string;
+  selectable?: boolean;
+  children?: React.ReactNode;
+};
+
+export default FooterProps;
