+++ workspaces/tables/src/private/types/HeaderProps.ts
@@ -1,0 +1,7 @@
+type HeaderProps = {
+  headerBackgroundColor: string;
+  selectable?: boolean;
+  children?: React.ReactNode;
+};
+
+export default HeaderProps;
