+++ workspaces/tables/src/private/types/HeaderTRProps.ts
@@ -1,0 +1,6 @@
+type HeaderTRProps = {
+  children: any;
+  reactTableProps?: any;
+};
+
+export default HeaderTRProps;
