+++ workspaces/tables/src/private/types/HiddenTHProps.ts
@@ -1,0 +1,6 @@
+type HiddenTHProps = {
+  style: React.CSSProperties;
+  children?: React.ReactNode;
+};
+
+export default HiddenTHProps;
