+++ workspaces/tables/src/private/types/MinimapProps.ts
@@ -1,0 +1,9 @@
+import type { MutableRefObject } from 'react';
+
+type MinimapProps = {
+  numberOfColumns: number;
+  tableRef: MutableRefObject<HTMLElement | null>;
+  footerRef: MutableRefObject<HTMLElement | null>;
+};
+
+export default MinimapProps;
