+++ workspaces/tables/src/private/types/RowSpanHeaderProps.ts
@@ -1,0 +1,6 @@
+type RowSpanHeaderProps = {
+  id: string;
+  value?: any;
+};
+
+export default RowSpanHeaderProps;
