+++ workspaces/tables/src/private/types/RowSubComponentProps.ts
@@ -1,0 +1,8 @@
+import type { Row } from 'react-table';
+
+type RowSubComponentProps<D extends object = {}> = {
+  row: Row<D>;
+  backgroundColor: string;
+};
+
+export default RowSubComponentProps;
