+++ workspaces/tables/src/private/types/TableLayout.ts
@@ -1,0 +1,3 @@
+type TableLayout = 'block';
+
+export default TableLayout;
