+++ workspaces/tables/src/private/types/TableVariant.ts
@@ -1,0 +1,3 @@
+type TableVariant = 'calendar';
+
+export default TableVariant;
