+++ workspaces/tables/src/private/types/TableWrapperProps.ts
@@ -1,0 +1,18 @@
+import type TableVariant from './TableVariant';
+
+type TableWrapperProps = {
+  reactTableProps: any;
+  isScrolledToLeft: boolean;
+  isScrolledToRight: boolean;
+  style?: React.CSSProperties;
+  className?: string;
+  variant?: TableVariant;
+  defaultRowBackgroundColor?: string;
+  evenRowBackgroundColor?: string;
+  headerBackgroundColor?: string;
+  children?: React.ReactNode;
+  tableRef?: any;
+  onScroll?: (tableRef: any) => any;
+};
+
+export default TableWrapperProps;
