--- workspaces/tables/src/styled/StyledTable.ts
@@ -1,14 +1,0 @@
-export const table =
-  'flex flex-col gap-1 w-full border-spacing-0 rounded-moon-s-sm p-1 bg-goku';
-export const thead = 'flex flex-col gap-1 sticky';
-export const tbody = 'flex flex-col gap-1 relative z-0';
-export const trHeader =
-  'flex gap-4 rounded-moon-s-sm text-bulma moon-text-14 px-4 sticky';
-export const trBody =
-  'flex gap-4 rounded-moon-s-sm bg-gohan transition-colors text-bulma moon-text-14 px-4';
-export const th = 'py-3 text-start font-normal';
-export const td = 'py-4 relative text-start ';
-export const tdDevider =
-  'after:content-[""] after:absolute after:top-[16px] after:bottom-[16px] after:w-px after:right-0 after:translate-x-[-50%] after:bg-beerus last:after:hidden';
-
-export default { table, thead, tbody, trHeader, trBody, th, td, tdDevider };
